backend/requirements.txt
Flask>=2.0.2
PyMySQL
Flask-SQLAlchemy
flask-marshmallow
marshmallow-sqlalchemy
google-api-python-client
pandas
flask-cors
pause
Flask-JWT-Extended
redis==3.5.0
celery